refactor: 플랜 생성 및 수정 요청 dto 클래스 형식 수정 (#15)
